Extract PageProjectionsProcessor from PageProcessor
